+++ CHANGELOG.rst
@@ -26,6 +26,9 @@ v34.12.0 (unreleased)
   Refine the CSS for the Resource viewer.
   https://github.com/aboutcode-org/scancode.io/pull/1692
 
+- Add "(No value detected)" for Copyright and Holder charts.
+  https://github.com/aboutcode-org/scancode.io/issues/1697
+
 v34.11.0 (2025-05-02)
 ---------------------